+++ gcovr/tests/different-function-lines-use-0/main.c
@@ -1,0 +1,9 @@
+#include <stdio.h>
+
+#include "func.h"
+
+int main(int argc, char* argv[]) {
+  foo(0);
+
+  return 0;
+}