+++ gcovr/tests/test_gcov_parser.py
@@ -27,7 +27,7 @@ from unittest import mock
 
 import pytest
 
-from ..gcov_parser import parse_coverage, UnknownLineType
+from ..gcov_parser import NegativeHits, parse_coverage, UnknownLineType
 from ..utils import configure_logging
 from ..workers import Workers
 from ..exclusions import ExclusionOptions, apply_all_exclusions
@@ -312,7 +312,7 @@ def test_gcov_8(capsys, sourcename):
     coverage, lines = parse_coverage(
         filename="tmp.cpp",
         lines=lines,
-        ignore_parse_errors=False,
+        ignore_parse_errors=None,
     )
 
     apply_all_exclusions(
@@ -350,7 +350,7 @@ def test_unknown_tags(caplog, ignore_errors):
         coverage, _ = parse_coverage(
             filename="foo.c",
             lines=lines,
-            ignore_parse_errors=ignore_errors,
+            ignore_parse_errors=["all"] if ignore_errors else None,
         )
         return coverage
 
@@ -392,7 +392,7 @@ def test_pathologic_codeline(caplog):
         parse_coverage(
             filename="foo.c",
             lines=lines,
-            ignore_parse_errors=False,
+            ignore_parse_errors=None,
         )
 
     messages = caplog.record_tuples
@@ -445,7 +445,7 @@ def test_exception_during_coverage_processing(caplog):
             parse_coverage(
                 lines,
                 filename="test.cpp",
-                ignore_parse_errors=False,
+                ignore_parse_errors=None,
             )
 
     # check that this is our exception
@@ -493,14 +493,14 @@ def test_trailing_function_tag():
     coverage, _ = parse_coverage(
         source.splitlines(),
         filename="test.cpp",
-        ignore_parse_errors=False,
+        ignore_parse_errors=None,
     )
 
     assert coverage.functions.keys() == {"example"}
     fcov = coverage.functions["example"]
-    assert fcov.lineno == 3  # previous lineno + 1
+    assert list(fcov.count.keys()) == [3]  # previous lineno + 1
     assert fcov.name == "example"
-    assert fcov.count == 17  # number of calls
+    assert fcov.count[3] == 17  # number of calls
 
 
 @pytest.mark.parametrize(
@@ -539,7 +539,7 @@ def test_branch_exclusion(flags):
     coverage, lines = parse_coverage(
         source.splitlines(),
         filename="example.cpp",
-        ignore_parse_errors=False,
+        ignore_parse_errors=None,
     )
 
     apply_all_exclusions(
@@ -556,6 +556,62 @@ def test_branch_exclusion(flags):
     }
 
     assert covered_branches == expected_covered_branches
+
+
+def test_negativ_branch_count():
+    """
+    A exception shall be raised.
+    """
+
+    source = textwrap.dedent(
+        """\
+          1: 1: normal line
+        branch 1 taken 80%
+          1: 2: } // line without apparent code
+        branch 2 taken -11234
+          1: 3: exception-only code
+        branch 3 taken 60% (throw)
+        """
+    )
+
+    with pytest.raises(NegativeHits):
+        parse_coverage(
+            source.splitlines(),
+            filename="example.cpp",
+            ignore_parse_errors=None,
+        )
+
+
+def test_negativ_branch_count_ignored():
+    """
+    A exception shall be raised.
+    """
+
+    source = textwrap.dedent(
+        """\
+          1: 1: normal line
+        branch 1 taken 80%
+          1: 2: } // line without apparent code
+        branch 2 taken -11234
+          1: 3: exception-only code
+        branch 3 taken 60% (throw)
+        """
+    )
+
+    coverage, lines = parse_coverage(
+        source.splitlines(),
+        filename="example.cpp",
+        ignore_parse_errors=["negative_hits.warn"],
+    )
+
+    covered_branches = {
+        branch
+        for line in coverage.lines.values()
+        for branch in line.branches.keys()
+        if line.branches[branch].is_covered
+    }
+
+    assert covered_branches == {1, 3}
 
 
 @pytest.mark.parametrize("flags", ["none", "exclude_internal_functions"])
@@ -579,7 +635,7 @@ def test_function_exclusion(flags):
     coverage, lines = parse_coverage(
         source.splitlines(),
         filename="example.cpp",
-        ignore_parse_errors=False,
+        ignore_parse_errors=None,
     )
 
     apply_all_exclusions(
@@ -610,14 +666,18 @@ def test_noncode_lines():
         lines: list,
         *,
         exclude_function_lines: bool = False,
+        exclude_noncode_lines: bool = False,
     ):
         coverage, source = parse_coverage(
             lines,
             filename="example.cpp",
-            ignore_parse_errors=False,
+            ignore_parse_errors=None,
         )
 
-        options = ExclusionOptions(exclude_function_lines=exclude_function_lines)
+        options = ExclusionOptions(
+            exclude_function_lines=exclude_function_lines,
+            exclude_noncode_lines=exclude_noncode_lines,
+        )
         apply_all_exclusions(coverage, lines=source, options=options)
 
         for line_data in coverage.lines.values():
@@ -659,7 +719,7 @@ def test_noncode_lines():
     assert get_line_status(["#####: 32:looks like code"]) == "normal:0"
 
     # Uncovered code that doesn't look like code: discard
-    assert get_line_status(["#####: 32:}"]) == "noncode"
+    assert get_line_status(["#####: 32:}"], exclude_noncode_lines=True) == "noncode"
 
 
 def check_and_raise(number, mutable, exc_raised, queue_full):